all of the data is now is separated files, cleaner look
