+++ index.html
@@ -98,6 +98,9 @@
     <script type="module" src="./scripts/home.js"></script>
     <script type="module" src="./scripts/placesFunctions.js"></script>
     <script type="module" src="./scripts/filteringFunctions.js"></script>
+    <script type="module" src="./scripts/restaurants.js"></script>
+    <script type="module" src="./scripts/bars.js"></script>
+    <script type="module" src="./scripts/filters.js"></script>
 </body>
 
 </html>
