+++ scripts/bars.js
@@ -1,0 +1,82 @@
+export const BarDataBase = [
+    {
+        name: 'Napoleone',
+        miniDescription: 'le QG des wilder, à boire et à manger',
+        specialities: 'cocktail vin biere tapas',
+        distance: '85 m',
+        priceRange: '€€-€€€',
+        adress: ' 1 rue baron',
+        tel: '02 53 78 40 61',
+        linkToMaps: 'https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2710.2721409788555!2d-1.5495075489104218!3d47.21125717905818!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x4805eeb05311c6ab%3A0x1a10a632301380c9!2s1+Rue+Baron%2C+44000+Nantes!5e0!3m2!1sfr!2sfr!4v1564840539599!5m2!1sfr!2sfr',
+        image: "./assets/Bars/Napoleone.png",
+        description: 'Cocktails créations, vins et bières locales, Bar a tapas'
+
+    },
+    {
+        name: 'Le lovecraft',
+        miniDescription: 'le QG des wilder, à boire et à manger',
+        specialities: 'biere artisanales vin cocktails',
+        distance: '400 m',
+        priceRange: '€',
+        adress: '32 bis Rue Fouré',
+        tel: '0228499855',
+        linkToMaps: '',
+        image: './assets/Bars/Lelovecraft.jpg',
+        description: 'le qg des wilder et aussi des buveurs! Pour les amateur de soft, le cocktail papaye est un incontournable'
+
+    },
+    {
+        name: 'Le lieu unique',
+        miniDescription: 'bar dans une ancienne biscuiterie',
+        specialities: 'bieres artisanales vins raffinés cocktails maison',
+        distance: '400 m',
+        priceRange: '€',
+        adress: '32 bis Rue Fouré',
+        tel: '0228499855',
+        linkToMaps: '',
+        image: './assets/Bars/barlu.png',
+        description: 'le qg des wilder et aussi des buveurs! Pour les amateur de soft, le cocktail papaye est un incontournable'
+
+    },
+    {
+        name: 'le bain rouge',
+        miniDescription: 'Lieu de spectacles et de tenue d’évènements',
+        specialities: 'bieres artisanales vins raffinés cocktails maison',
+        distance: '400 m',
+        priceRange: '€',
+        adress: '32 bis Rue Fouré',
+        tel: '0228499855',
+        linkToMaps: '',
+        image: './assets/Bars/bainrouge.jpg',
+        description: 'le qg des wilder et aussi des buveurs! Pour les amateur de soft, le cocktail papaye est un incontournable'
+
+    },
+    {
+        name: 'Coco bandit',
+        miniDescription: 'le QG des wilder, à boire et à manger',
+        specialities: 'cocktail',
+        distance: '400 m',
+        priceRange: '€',
+        adress: '15 Rue Crucy ',
+        tel: '0640784675',
+        linkToMaps: '',
+        image: './assets/Bars/cocobandit.jpg',
+        description: 'le qg des wilder et aussi des buveurs! Pour les amateur de soft, le cocktail papaye est un incontournable'
+
+    },
+    {
+        name: 'Delirium cafe',
+        miniDescription: 'sélection de bières pression et soirées musicales ou sportives.',
+        specialities: 'bieres artisanales vins raffinés cocktails maison',
+        distance: '400 m',
+        priceRange: '€',
+        adress: '32 bis Rue Fouré',
+        tel: '0228499855',
+        linkToMaps: '',
+        image: './assets/Bars/Delirium-Cafe.jpg',
+        description: 'le qg des wilder et aussi des buveurs! Pour les amateur de soft, le cocktail papaye est un incontournable'
+
+    }
+]
+
+
