+++ scripts/filters.js
@@ -1,0 +1,89 @@
+export const primaryFoodFilters = [
+    {
+        name: 'italien',
+        icon: './icons/pizza.svg',
+        iconPink: './icons/pizza_pink.svg',
+        active: false
+    },
+    {
+        name: 'sandwich',
+        icon: './icons/sandwich.svg',
+        iconPink: './icons/sandwich_pink.svg',
+        active: false
+    },
+    {
+        name: 'asiatique',
+        icon: './icons/ramen.svg',
+        iconPink: './icons/ramen_pink.svg',
+        active: false
+    },
+    {
+        name: 'burger',
+        icon: './icons/burger.svg',
+        iconPink: './icons/burger_pink.svg',
+        active: false
+    },
+    {
+        name: 'dessert',
+        icon: './icons/cupcake.svg',
+        iconPink: './icons/cupcake_pink.svg',
+        active: false
+    },
+    {
+        name: 'salade',
+        icon: './icons/salade.svg',
+        iconPink: './icons/salade_pink.svg',
+        active: false
+    }
+];
+
+
+export const secondaryFoodFilters = [
+    {
+        name: 'vegan',
+        icon: './icons/vegan.svg',
+        iconPink: './icons/vegan_pink.svg',
+        active: false
+    },
+    {
+        name: 'halal',
+        icon: './icons/halal.svg',
+        iconPink: './icons/halal_pink.svg',
+        active: false
+    },
+    {
+        name: 'sans gluten',
+        icon: './icons/glutenfree.svg',
+        iconPink: './icons/glutenfree_pink.svg',
+        active: false
+    }
+];
+
+
+export const primaryDrinkFilters = [
+    {
+        name: 'biere',
+        icon: './icons/beer.svg',
+        iconPink: './icons/beer_pink.svg',
+        active: false
+    },
+    {
+        name: 'cafe',
+        icon: './icons/coffee.svg',
+        iconPink: './icons/coffee_pink.svg',
+        active: false
+    },
+    {
+        name: 'vin',
+        icon: './icons/wine.svg',
+        iconPink: './icons/wine_pink.svg',
+        active: false
+    },
+    {
+        name: 'cocktails',
+        icon: './icons/cocktail.svg',
+        iconPink: './icons/cocktail_pink.svg',
+        active: false
+    }
+
+];
